requirements.txt
aniso8601==9.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
Faker==18.11.2
Flask==1.1.2
Flask-RESTful==0.3.8
idna==3.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pymongo==3.10.1
python-dateutil==2.8.2
python-dotenv
pytz==2023.3
requests==2.31.0
six==1.16.0
urllib3==2.0.3
Werkzeug==0.16.1
